# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/admin/filters": {
      "get": {
        "operationId": "AdminController_getFilters",
        "summary": "Get filter options",
        "description": "Get all available filter options (stores, regions, cities, categories, brands, manufacturers) with their IDs. Use this to get IDs needed for product creation.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Filter options with IDs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "stores": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          }
                        }
                      }
                    },
                    "categories": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          }
                        }
                      }
                    },
                    "brands": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          }
                        }
                      }
                    },
                    "manufacturers": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/admin/categories": {
      "get": {
        "operationId": "AdminController_getCategories",
        "summary": "Get all categories",
        "description": "Get list of all categories with their IDs. Use these IDs when creating products via POST /products endpoint.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of categories",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "example": "4158f662-9e15-42ef-9c4e-345b9465693c"
                      },
                      "name": {
                        "type": "string",
                        "example": "Beverages"
                      },
                      "parentId": {
                        "type": "string",
                        "nullable": true,
                        "example": null
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/admin/brands": {
      "get": {
        "operationId": "AdminController_getBrands",
        "summary": "Get all brands",
        "description": "Get list of all brands with their IDs and manufacturer info. Use these IDs when creating products via POST /products endpoint.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of brands",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "example": "4158f662-9e15-42ef-9c4e-345b9465693c"
                      },
                      "name": {
                        "type": "string",
                        "example": "Coca-Cola"
                      },
                      "manufacturerId": {
                        "type": "string",
                        "example": "4158f662-9e15-42ef-9c4e-345b9465693c"
                      },
                      "manufacturerName": {
                        "type": "string",
                        "example": "The Coca-Cola Company"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/admin/manufacturers": {
      "get": {
        "operationId": "AdminController_getManufacturers",
        "summary": "Get all manufacturers",
        "description": "Get list of all manufacturers with their IDs. Use these IDs when creating products via POST /products endpoint.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of manufacturers",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string",
                        "example": "4158f662-9e15-42ef-9c4e-345b9465693c"
                      },
                      "name": {
                        "type": "string",
                        "example": "The Coca-Cola Company"
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/admin/analytics/summary": {
      "get": {
        "operationId": "AdminController_getSummary",
        "summary": "Get dashboard summary",
        "description": "Get summary metrics for admin dashboard (total sales, invoices, AOV, etc.)",
        "parameters": [
          {
            "name": "storeId",
            "required": false,
            "in": "query",
            "description": "Filter by store ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "description": "End date (YYYY-MM-DD)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "description": "Start date (YYYY-MM-DD)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Summary metrics"
          }
        },
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/admin/analytics/sales-trend": {
      "get": {
        "operationId": "AdminController_getSalesTrend",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/admin/analytics/brand-distribution": {
      "get": {
        "operationId": "AdminController_getBrandDistribution",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/admin/analytics/market-share": {
      "get": {
        "operationId": "AdminController_getMarketShare",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/admin/analytics/brands": {
      "get": {
        "operationId": "AdminController_getBrandsAnalytics",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/admin/analytics/top-skus": {
      "get": {
        "operationId": "AdminController_getTopSkus",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Meri Dukaan POS + Admin Analytics API",
    "description": "Complete API documentation for Meri Dukaan Point of Sale system with Admin Analytics",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
...
  }
}